Move port selection into mission main form locations
