+++ frontend/cypress/e2e/side_window/mission_form/land_control.spec.ts
@@ -23,9 +23,6 @@ context('Side Window > Mission Form > Land Control', () => {
     // Date et heure du contrôle
     // TODO Add this test.
 
-    // Lieu du contrôle
-    cy.fill('Lieu du contrôle', 'Free Port')
-
     // Obligations déclaratives et autorisations de pêche
     cy.fill('Bonne émission VMS', 'Oui')
     cy.fill('Bonne émission AIS', 'Non')
@@ -88,10 +85,7 @@ context('Side Window > Mission Form > Land Control', () => {
     // -------------------------------------------------------------------------
     // Request
 
-    cy.intercept('POST', '/bff/v1/mission_actions', {
-      // TODO This should be removed once the API works as expected.
-      statusCode: 201
-    }).as('createMissionAction')
+    cy.intercept('POST', '/bff/v1/mission_actions').as('createMissionAction')
 
     cy.clickButton('Enregistrer et clôturer')
 
@@ -141,8 +135,6 @@ context('Side Window > Mission Form > Land Control', () => {
         otherInfractions: [
           { comments: 'Une observation sur l’infraction autre.', infractionType: 'WITHOUT_RECORD', natinf: 27689 }
         ],
-        portLocode: 'AEFRP',
-        portName: 'Free Port',
         segments: [
           {
             faoAreas: ['37.1', '37.2', '37.3', '27.8.a', '27.8.b', '27.7.h', '27.7.e', '27.7.d'],
