+++ frontend/cypress/e2e/side_window/mission_form/observation.spec.ts
@@ -24,10 +24,7 @@ context('Side Window > Mission Form > Observation', () => {
     // -------------------------------------------------------------------------
     // Request
 
-    cy.intercept('POST', '/bff/v1/mission_actions', {
-      // TODO This should be removed once the API works as expected.
-      statusCode: 201
-    }).as('createMissionAction')
+    cy.intercept('POST', '/bff/v1/mission_actions').as('createMissionAction')
 
     cy.clickButton('Enregistrer et clôturer')
