+++ frontend/src/features/SideWindow/MissionForm/MainForm/index.tsx
@@ -13,8 +13,8 @@ import { useCallback, useMemo, useRef } from 'react'
 import styled from 'styled-components'
 
 import { MISSION_TYPES_AS_OPTIONS } from './constants'
+import { FormikLocationPicker } from './FormikLocationPicker'
 import { FormikMultiControlUnitPicker } from './FormikMultiControlUnitPicker'
-import { FormikMultiZonePicker } from './FormikMultiZonePicker'
 import { BOOLEAN_AS_OPTIONS } from '../../../../constants'
 import { useNewWindow } from '../../../../ui/NewWindow'
 import { FormBody, FormBodyInnerWrapper } from '../shared/FormBody'
@@ -70,25 +70,16 @@ export function MainForm({ initialValues, onChange }: MainFormProps) {
                 options={MISSION_TYPES_AS_OPTIONS}
               />
 
-              {/* TODO Fix that in Monitor UI: */}
-              {/* Re-enabling a checkbox that has been disabled should set the related FormValues prop
-                  to a boolean matching the checkbox `checked` state. */}
               <IsUnderJdpFormikCheckbox isUndefinedWhenDisabled label="Mission sous JDP" name="isUnderJdp" />
             </MultiCheckColumns>
 
-            <FormikMultiRadio
-              isInline
-              label="Ordre de mission"
-              name="hasOrder"
-              // TODO Allow more Monitor UI `Option` types.
-              options={BOOLEAN_AS_OPTIONS}
-            />
+            <FormikMultiRadio isInline label="Ordre de mission" name="hasOrder" options={BOOLEAN_AS_OPTIONS} />
           </CustomFormBodyInnerWrapper>
 
           <FormikMultiControlUnitPicker name="controlUnits" />
 
           <CustomFormBodyInnerWrapper>
-            <FormikMultiZonePicker name="geom" />
+            <FormikLocationPicker />
 
             <RelatedFieldGroupWrapper>
               <FormikTextarea label="CACEM : orientations, observations" name="observationsCacem" />
